decouple field template logic
