+++ app/src/composables/use-template/index.ts
@@ -1,0 +1,4 @@
+import {useTemplate} from './use-template';
+
+export default useTemplate;
+export { useTemplate };
